refactor permission for remote session
