+++ core-ai-cli/src/main/java/ai/core/cli/CliApp.java
@@ -17,7 +17,7 @@ import ai.core.cli.ui.TerminalUI;
 import ai.core.session.FileSessionPersistence;
 import ai.core.session.SessionManager;
 import ai.core.session.SessionPersistence.SessionInfo;
-import ai.core.session.ToolPermissionStore;
+import ai.core.session.FileToolPermissionStore;
 
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -73,7 +73,7 @@ public class CliApp {
             currentSessionId = "cli-" + LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyyMMdd-HHmmss"));
         }
 
-        var permissionStore = new ToolPermissionStore(workspace.resolve(".core-ai").resolve("tool-permissions.txt"));
+        var permissionStore = new FileToolPermissionStore(workspace.resolve(".core-ai").resolve("tool-permissions.txt"));
         var noteMemory = new LocalFileMemoryProvider(workspace);
         var modelRegistry = new ModelRegistry(result.llmProviders, props);
